+++ NotationController.h
@@ -126,7 +126,7 @@ typedef struct _NoteCatalogEntry {
 - (void)databaseEncryptionSettingsChanged;
 - (void)databaseSettingsChangedFromOldFormat:(NSInteger)oldFormat;
 
-- (int)currentNoteStorageFormat;
+- (NSInteger)currentNoteStorageFormat;
 - (void)synchronizeNoteChanges:(NSTimer*)timer;
 
 - (void)updateDateStringsIfNecessary;
@@ -168,7 +168,7 @@ typedef struct _NoteCatalogEntry {
 - (BOOL)filterNotesFromUTF8String:(const char*)searchString forceUncached:(BOOL)forceUncached;
 - (NSUInteger)preferredSelectedNoteIndex;
 - (NSArray*)noteTitlesPrefixedByString:(NSString*)prefixString indexOfSelectedItem:(NSInteger *)anIndex;
-- (NoteObject*)noteObjectAtFilteredIndex:(int)noteIndex;
+- (NoteObject*)noteObjectAtFilteredIndex:(NSUInteger)noteIndex;
 - (NSArray*)notesAtIndexes:(NSIndexSet*)indexSet;
 - (NSIndexSet*)indexesOfNotes:(NSArray*)noteSet;
 - (NSUInteger)indexInFilteredListForNoteIdenticalTo:(NoteObject*)note;
