+++ PTHotKeys/PTKeyComboPanel.m
@@ -97,7 +97,7 @@ static id _sharedKeyComboPanel = nil;
 }
 
 - (void)runModalForHotKey: (PTHotKey*)hotKey {
-	int resultCode;
+	NSInteger resultCode;
     
     [self setKeyCombo: [hotKey keyCombo]];
 	[self setKeyBindingName: [hotKey name]];
